Add Interfaces for Services and Repositories
