+++ src/main/java/com/adamant/locationservice/service/CouchDBService.java
@@ -1,116 +1,19 @@
 package com.adamant.locationservice.service;
 
 import com.adamant.locationservice.entity.couch_db.CouchDBFindAllResponse;
-import com.adamant.locationservice.entity.couch_db.CouchDBReplicationResponse;
-import com.adamant.locationservice.entity.couch_db.CouchDBSaveResponse;
-import com.adamant.locationservice.entity.couch_db.CouchDBUpdateResponse;
 import com.adamant.locationservice.entity.user_management.CouchDocument;
-import com.adamant.locationservice.repository.CouchDBRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.util.ObjectUtils;
 
-import java.util.Locale;
+public interface CouchDBService {
 
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class CouchDBService {
+    <T extends CouchDocument> boolean save(T t);
 
-    private final CouchDBRepository couchDBRepository;
+    <T> boolean update(T t);
 
-    /**
-     * Save object in the database.
-     *
-     * @param t   - Object to save
-     * @param <T> - Type of the object to save.
-     * @return Is the save success
-     */
-    public <T extends CouchDocument> boolean save(T t) {
-        t.setType(t.getClass().getSimpleName().toLowerCase(Locale.ENGLISH));
-        CouchDBSaveResponse response = couchDBRepository.save(t);
+    <T> T find(Class<T> classType, String id);
 
-        if (ObjectUtils.isEmpty(response.getError()) && !ObjectUtils.isEmpty(response.getId())) {
-            return true;
-        }
-        log.error("Error occurred in saving the object: {} error: {} reason: {}", t.toString(), response.getError(), response.getReason());
-        return false;
-    }
+    <T> CouchDBFindAllResponse<T> findAll(Class<T> classType, int rowsPerPage, String param);
 
-    /**
-     * Update object in the database.
-     *
-     * @param t   - Object to update
-     * @param <T> - Type of the object to update.
-     * @return Is the update success
-     */
-    public <T> boolean update(T t) {
-        CouchDBUpdateResponse response = couchDBRepository.update(t);
+    boolean remove(String id, String rev);
 
-        if (ObjectUtils.isEmpty(response.getError()) && !ObjectUtils.isEmpty(response.getId())) {
-            return true;
-        }
-        log.error("Error occurred in updating the object: {} error: {} reason: {}", t.toString(), response.getError(), response.getReason());
-        return false;
-    }
-
-    /**
-     * Find an object with given class type with the document id.
-     *
-     * @param classType - The class of type T
-     * @param id        - Document id
-     * @param <T>       - Object type
-     * @return An object of type T.
-     */
-    public <T> T find(Class<T> classType, String id) {
-        return couchDBRepository.find(classType, id);
-    }
-
-    /**
-     * @param classType   - The class of type T
-     * @param rowsPerPage - rows per page
-     * @param param       - param to search page
-     * @param <T>         - Object type
-     * @return An object of type CouchDBFindAllResponse
-     */
-    public <T> CouchDBFindAllResponse<T> findAll(Class<T> classType, int rowsPerPage, String param) {
-        return couchDBRepository.findAll(classType, classType.getSimpleName().toLowerCase(Locale.ENGLISH), rowsPerPage, param);
-    }
-
-    /**
-     * Remove a given object from the database.
-     *
-     * @param id  - id
-     * @param rev - revision
-     * @return Is the removal success
-     */
-    public boolean remove(String id, String rev) {
-        CouchDBSaveResponse response = couchDBRepository.remove(id, rev);
-
-        if (ObjectUtils.isEmpty(response.getError()) && !ObjectUtils.isEmpty(response.getId())) {
-            return true;
-        }
-        log.error("Error occurred in removing the object with id: {} and rev: {} error: {} reason: {}", id, rev, response.getError(), response.getReason());
-        return false;
-    }
-
-    /**
-     * Trigger a replication between source database and a target.
-     *
-     * @param source - Source database
-     * @param target - Target database
-     * @return Is the replication success
-     */
-    public boolean replicate(String source, String target) {
-
-        CouchDBReplicationResponse response = couchDBRepository.replicate(source, target);
-
-        if (response.isOk()) {
-            return true;
-        }
-        log.error("Error occurred in replicating with target: {}", target);
-        return false;
-
-    }
+    boolean replicate(String source, String target);
 }
